Create 'logic/account' + rename some stuff
